Added management for second run with Tsan

+++ lib/rtl/swordrt_rtl.cc
@@ -28,30 +28,74 @@
 		if((access >= (size_t) stack) &&						\
 				(access < (size_t) stack + stacksize))			\
 				return;
-#define SAVE_ACCESS(size, type)																			\
-		std::unordered_map<uint64_t, AccessInfo>::iterator item = accesses.find(hash); 					\
-		if(item == accesses.end()) {																	\
-			accesses.insert(std::make_pair(hash, AccessInfo(access, access, UINT_MAX, 0, size, type, pc)));	\
-		} else {																						\
-			if(access < item->second.address) {															\
-				item->second.address = access;															\
-				item->second.count++;																	\
-				int64_t diff = item->second.prev_address - access;										\
-				unsigned stride = diff / (1 << size);													\
-				if(stride < item->second.stride)														\
-					item->second.stride = stride;														\
-				INFO(std::cout, "PATTERN:" << tid << "," << hash << ","<< access << "," << item->second.prev_address << "," << item->second.stride << "," << (1 << size) << "," << pc);			\
-				item->second.prev_address = access;														\
-			} else if(((access - item->second.address) / (1 << size)) >= item->second.count + 1) {			\
-				item->second.count++;																	\
-				int64_t diff = access - item->second.prev_address;										\
-				unsigned stride = diff / (1 << size);													\
-				if(stride < item->second.stride)														\
-					item->second.stride = stride;														\
-				INFO(std::cout, "PATTERN:" << tid << "," << hash << ","<< access << "," << item->second.prev_address << "," << item->second.stride << "," << (1 << size) << "," << pc);			\
-				item->second.prev_address = access;														\
-			}																							\
+#define TSAN_CHECK(name)																											\
+		if(tsan_enabled) { 																											\
+			std::set<std::pair<uint64_t, uint64_t>>::iterator it = pairs_tsan_checks.find(std::make_pair(__swordrt_hash__, hash));  \
+			if(it != pairs_tsan_checks.end())																						\
+				__tsan_ ## name(addr);																								\
+			return;  																												\
+  	    }
+#define SAVE_ACCESS(size, type)															\
+		uint64_t diff;																	\
+		std::unordered_map<uint64_t, AccessInfo>::iterator item = accesses.find(hash); 	\
+		if(item == accesses.end()) {													\
+			accesses.insert(std::make_pair(hash, AccessInfo(access, access, UINT_MAX,	\
+							0, ULLONG_MAX, size, type, pc)));							\
+		} else {																		\
+			switch(item->second.count) {												\
+			case 0:																		\
+				break;																	\
+			/* Second access */															\
+			case ULLONG_MAX:															\
+				if(access < item->second.address) {										\
+					diff = item->second.prev_address - access;							\
+					item->second.address = access;										\
+					item->second.count = 1;												\
+					item->second.diff = diff;											\
+				    item->second.stride = 1;											\
+				    item->second.prev_address = access;									\
+				} else if(access > item->second.address) {								\
+					diff = access - item->second.prev_address;							\
+					item->second.count = 1;												\
+					item->second.diff = diff;											\
+				    item->second.stride = 1;											\
+				    item->second.prev_address = access;									\
+				} else {																\
+					item->second.count = 0;												\
+				}																		\
+				break;																	\
+			default:																	\
+				if(access < item->second.prev_address) {								\
+					diff = item->second.prev_address - access;							\
+					if(diff == item->second.diff) {										\
+						item->second.address = access;									\
+						item->second.count++;											\
+					    item->second.stride = 1;										\
+					    item->second.prev_address = access;								\
+					} else {															\
+						/* Check with Tsan */											\
+						tsan_checks.insert(hash);										\
+					}																	\
+				} else if(access > item->second.prev_address) {							\
+					diff = access - item->second.prev_address;							\
+					if(diff == item->second.diff) {										\
+						item->second.count++;											\
+					    item->second.stride = 1;										\
+					    item->second.prev_address = access;								\
+					} else {															\
+						/* Check with Tsan */											\
+						/* DEBUG(std::cout, "Check tsan" << pc); */						\
+						tsan_checks.insert(hash);										\
+					}																	\
+				} else {																\
+					item->second.count = 0;												\
+				}																		\
+				break;																	\
+			}																			\
 		}
+
+//INFO(std::cout, "PATTERN:" << tid << "," << hash << ","<< access << "," << item->second.prev_address << "," << item->second.stride <<	"," << (1 << size) << "," << pc);
+
 #else
 #define GET_STACK	 										\
 		pthread_t self = pthread_self(); 					\
@@ -164,9 +208,14 @@ static void on_ompt_event_barrier_begin(ompt_parallel_id_t parallel_id,
 
 	__swordrt_barrier__++;
 
+	for(std::set<uint64_t>::iterator it = tsan_checks.begin(); it != tsan_checks.end(); ++it)
+		accesses.erase(*it);
+
 #if defined(TLS) || defined(NOTLS)
 	oss << "DATA_BEGIN[" << std::dec << parallel_id << "," << tid << "," << omp_get_thread_num() << ":" << omp_get_num_threads() << "," << __swordrt_barrier__ << "]\n";
 	for (std::unordered_map<uint64_t, AccessInfo>::iterator it = accesses.begin(); it != accesses.end(); ++it) {
+		if(it->second.count == ULLONG_MAX)
+			it->second.count = 0;
 		oss << "DATA[" << std::dec << it->first << "," << std::hex << "0x" << it->second.address << "," << std::dec << it->second.count << "," << it->second.size << "," << it->second.type << "," << "0x" << std::hex << it->second.pc << "," << std::dec << it->second.stride << "]\n";
 	}
 	oss << "DATA_END[" << std::dec << parallel_id << "," << tid << "," << omp_get_thread_num() << ":" << omp_get_num_threads() << "," << __swordrt_barrier__ << "]\n";
@@ -182,34 +231,67 @@ static void on_ompt_event_barrier_begin(ompt_parallel_id_t parallel_id,
 	threadInfo[tid - 1].accesses.clear();
 #endif
 
+	smtx.lock();
+	// total_tsan_checks.insert(tsan_checks.begin(), tsan_checks.end());
+	if(tsan_checks.size() > 0) {
+		setdatafile.open(std::string(ARCHER_DATA) + "/tsanchecks", std::ofstream::app);
+		for(std::set<uint64_t>::iterator it = tsan_checks.begin(); it != tsan_checks.end(); ++it)
+			DATA(setdatafile, __swordrt_hash__ << "," << *it << "\n");
+		tsan_checks.clear();
+		setdatafile.close();
+	}
+	smtx.unlock();
 }
 
 static void ompt_initialize_fn(ompt_function_lookup_t lookup,
 		const char *runtime_version,
 		unsigned int ompt_version) {
-	DEBUG(std::cout, "OMPT Initizialization: Runtime Version: " << std::dec << runtime_version << ", OMPT Version: " << std::dec << ompt_version);
+	FILE *fp;
+	std::string filename = std::string(ARCHER_DATA) + "/tsanchecks";
+	if((fp = fopen(filename.c_str(), "r")) != NULL) {
+		pairs_tsan_checks.clear();
+		uint64_t hash1, hash2;
+		while (true) {
+			int ret = fscanf(fp, "%lu,%lu", &hash1, &hash2);
+			if(ret == 2) {
+				entry_tsan_checks.insert(hash1);
+				pairs_tsan_checks.insert(std::make_pair(hash1, hash2));
+			} else if(errno != 0) {
+				perror("scanf:");
+				break;
+			} else if(ret == EOF) {
+				break;
+			} else {
+				DEBUG(std::cerr, "No match.");
+			}
+		}
+	} else {
+		std::string str = "rm -rf " + std::string(ARCHER_DATA);
+		system(str.c_str());
+		str = "mkdir " + std::string(ARCHER_DATA);
+		system(str.c_str());
+	}
 
-	std::string str = "rm -rf " + std::string(ARCHER_DATA);
-	system(str.c_str());
-	str = "mkdir " + std::string(ARCHER_DATA);
-	system(str.c_str());
+	if(!tsan_enabled) {
+		DEBUG(std::cout, "OMPT Initizialization: Runtime Version: " << std::dec << runtime_version << ", OMPT Version: " << std::dec << ompt_version);
 
-	ompt_set_callback_t ompt_set_callback = (ompt_set_callback_t) lookup("ompt_set_callback");
-	ompt_get_thread_id = (ompt_get_thread_id_t) lookup("ompt_get_thread_id");
-	ompt_get_parallel_id = (ompt_get_parallel_id_t) lookup("ompt_get_parallel_id");
+		ompt_set_callback_t ompt_set_callback = (ompt_set_callback_t) lookup("ompt_set_callback");
+		ompt_get_thread_id = (ompt_get_thread_id_t) lookup("ompt_get_thread_id");
+		ompt_get_parallel_id = (ompt_get_parallel_id_t) lookup("ompt_get_parallel_id");
 
-	ompt_set_callback(ompt_event_thread_begin,
-			(ompt_callback_t) &on_ompt_event_thread_begin);
-	ompt_set_callback(ompt_event_parallel_begin,
-			(ompt_callback_t) &on_ompt_event_parallel_begin);
-	ompt_set_callback(ompt_event_parallel_end,
-			(ompt_callback_t) &on_ompt_event_parallel_end);
-	ompt_set_callback(ompt_event_acquired_critical,
-			(ompt_callback_t) &on_acquired_critical);
-	ompt_set_callback(ompt_event_release_critical,
-			(ompt_callback_t) &on_release_critical);
-	ompt_set_callback(ompt_event_barrier_begin,
-			(ompt_callback_t) &on_ompt_event_barrier_begin);
+		ompt_set_callback(ompt_event_thread_begin,
+				(ompt_callback_t) &on_ompt_event_thread_begin);
+		ompt_set_callback(ompt_event_parallel_begin,
+				(ompt_callback_t) &on_ompt_event_parallel_begin);
+		ompt_set_callback(ompt_event_parallel_end,
+				(ompt_callback_t) &on_ompt_event_parallel_end);
+		ompt_set_callback(ompt_event_acquired_critical,
+				(ompt_callback_t) &on_acquired_critical);
+		ompt_set_callback(ompt_event_release_critical,
+				(ompt_callback_t) &on_release_critical);
+		ompt_set_callback(ompt_event_barrier_begin,
+				(ompt_callback_t) &on_ompt_event_barrier_begin);
+	}
 }
 
 ompt_initialize_t ompt_tool(void) {
